requirements.txt
feedparser==6.0.11
readability-lxml==0.8.1
requests==2.31.0
google-generativeai==0.3.2
python-dotenv==1.0.0
lxml==5.1.0
beautifulsoup4==4.12.3

